version final

+++ src/app/front-admin/CRUD/competitions/delete-competitions/delete-competitions.component.ts
@@ -2,6 +2,7 @@ import { Component, Input, input } from '@angular/core';
 import Swal from 'sweetalert2';
 import { CompitetionServiceService } from '../../../../service/compitetion-service.service'; 
 import { ComputationService } from '../../../../service/shared/computation.service';
+import { TokenStorageServiceService } from '../../../../service/token-storage-service.service';
 
 @Component({
   selector: 'app-delete-competitions',
@@ -12,36 +13,48 @@ import { ComputationService } from '../../../../service/shared/computation.servi
 export class DeleteCompetitionsComponent {
   @Input() competition: any;
 
-  constructor(private compitetionServiceService: CompitetionServiceService, private computationService: ComputationService) { }
+  constructor(private compitetionServiceService: CompitetionServiceService, private computationService: ComputationService,
+    private tokenStorageService: TokenStorageServiceService
+  ) { }
 
   deleteCompetition() {
-    Swal.fire({
-      title: 'Are you sure?',
-      text: `Do you really want to delete competition with Code: ${this.competition.code}?`,
-      icon: 'warning',
-      showCancelButton: true,
-      confirmButtonColor: '#d33',
-      cancelButtonColor: '#3085d6',
-      confirmButtonText: 'Yes, delete it!',
-      cancelButtonText: 'Cancel',
-    }).then((result) => {
-      if (result.isConfirmed) {
-        this.compitetionServiceService.deleteCompetition(this.competition.id)
-        .subscribe(
-          (response)=>{
-            Swal.fire('Deleted!', 'The competition has been deleted.', 'success');
-            this.computationService.triggerRefresh();
-          },
-          (error)=>{
-            console.error('Error deleting competition:', error);
+    if(this.tokenStorageService.getSub() === 'jury_test'){
+      Swal.fire({
+        icon: 'error',
+        title: 'You are not allowed to delete competition',
+        showConfirmButton: false,
+        timer: 1500
+      });
+    }else{
+      Swal.fire({
+          title: 'Are you sure?',
+          text: `Do you really want to delete competition with Code: ${this.competition.code}?`,
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#d33',
+          cancelButtonColor: '#3085d6',
+          confirmButtonText: 'Yes, delete it!',
+          cancelButtonText: 'Cancel',
+        }).then((result) => {
+          if (result.isConfirmed) {
+            this.compitetionServiceService.deleteCompetition(this.competition.id)
+            .subscribe(
+              (response)=>{
+                Swal.fire('Deleted!', 'The competition has been deleted.', 'success');
+                this.computationService.triggerRefresh();
+              },
+              (error)=>{
+                console.error('Error deleting competition:', error);
+                
+                Swal.fire('Eroor!', 'Error on deletig competition.', 'error');
+              }
+            )
             
-            Swal.fire('Eroor!', 'Error on deletig competition.', 'error');
+          } else if (result.dismiss === Swal.DismissReason.cancel) {
+            Swal.fire('Cancelled', 'The competition is safe :)', 'info');
           }
-        )
-        
-      } else if (result.dismiss === Swal.DismissReason.cancel) {
-        Swal.fire('Cancelled', 'The competition is safe :)', 'info');
-      }
-    });
+        });
+    }
+ 
   }
 }
